add: splitting code  into multilayers - generateUploadUrl
